% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ats.R
\name{ats}
\alias{ats}
\title{Compute index and probability of stable threshold using automatic threshold selection}
\usage{
ats(s, type = "index")
}
\arguments{
\item{s}{A stabsel object}

\item{type}{Either "index" or "prob" to return index of last selected variable, or threshold for deeming variables stable}
}
\value{
Index or probability of the last selected threshold
}
\description{
Compute index and probability of stable threshold using automatic threshold selection
}
