% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stabs.R
\name{plot_stabs}
\alias{plot_stabs}
\title{Plot stabsel object}
\usage{
plot_stabs(s, which = 1, threshold = "ATS", top = ifelse(s$p < 10, s$p, 10))
}
\arguments{
\item{s}{A stabsel object}

\item{which}{Either 1 to plot variable probability, or 2 to plot likelihood function}

\item{threshold}{Show selected MTS or ATS variables. Either "ATS" or "MTS"}

\item{top}{A number plotting the top number of variables. Use "all" to plot all selected variables}
}
\value{
Variable Probability or Likelihood Plot
}
\description{
Plot stabsel object
}
\examples{
library(mhstability)
y = mtcars[,2]
X = mtcars[,3:ncol(mtcars)]
res = stabs::stabsel(x = X, y = y, B = 50, fitfun = stabs::lars.lasso, PFER = 2, cutoff = 0.6, sampling.type = "SS", assumption = "r-concave")
mhstability::plot_stabs(res, which = 1)
mhstability::plot_stabs(res, which = 2)
}
\references{
Hofner, B. and Hothorn, T. (2021). stabs: Stability Selection with Error Control.

Meinshausen, N. and Buehlmann, P. (2009). Stability Selection. arXiv:0809.2932 [stat]

Shah, R. D. and Samworth, R. J. (2013). Variable selection with error control:
another look at stability selection. Journal of the Royal Statistical Society. Series B (Statistical
Methodology), 75(1):55–80. Publisher: Wiley.

Zhu, M. and Ghodsi, A. (2006). Automatic dimensionality selection from the scree plot via the use of profile likelihood.
Computational Statistics & Data Analysis, 51(2):918–930.
}
