% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stabs.R
\name{plot_stabs}
\alias{plot_stabs}
\title{Plot stabsel object}
\usage{
plot_stabs(s, which = 1, threshold = "ATS", top = 10)
}
\arguments{
\item{s}{A stabsel object}

\item{which}{Either 1 to plot variable probability, or 2 to plot likelihood function}

\item{threshold}{Show selected MTS or ATS variables. Either "ATS" or "MTS"}

\item{top}{A number plotting the top number of variables. Use "all" to plot all selected variables}
}
\value{
Variable Probability or Likelihood Plot
}
\description{
Plot stabsel object
}
\examples{
plot_stabs(s, which = 2, threshold = "MTS", top = "all")
plot_stabs(s, which = 1, threshold = "ATS", top = 5)
}
